requirements.txt
absl-py==0.11.0
aiofiles==0.6.0
aiohttp==3.7.3
aniso8601==8.1.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
blinker==1.4
Bottleneck==1.3.2
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
fastai==1.0.61
fastprogress==1.0.0
Flask==1.1.2
Flask-DebugToolbar==0.11.0
Flask-RESTful==0.3.8
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h11==0.11.0
h5py==3.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
numexpr==2.7.1
numpy==1.19.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.5
Pillow @ file:///C:/ci/pillow_1603823068645/work
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-testadapter==0.3.0
requests-toolbelt==0.9.1
rsa==4.6
scipy==1.5.4
six @ file:///C:/ci/six_1605187374963/work
soupsieve==2.1
starlette==0.14.1
tb-nightly==2.5.0a20201219
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
termcolor==1.1.0
tf-estimator-nightly==2.5.0.dev2020121901
tf-nightly==2.5.0.dev20201217
torch==1.7.1
torchvision==0.8.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3==1.26.2
uvicorn==0.13.2
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
yarl==1.6.3
